+++ pkgs/ffigen/lib/src/header_parser/utils.dart
@@ -6,6 +6,7 @@ import 'dart:ffi';
 
 import 'package:ffi/ffi.dart';
 import 'package:logging/logging.dart';
+import 'package:pub_semver/pub_semver.dart';
 
 import '../code_generator.dart';
 import '../config_provider/config_types.dart';
@@ -383,15 +384,14 @@ extension CXStringExt on clang_types.CXString {
 }
 
 extension CXVersionExt on clang_types.CXVersion {
-  VersionTriple? get triple {
+  Version? get triple {
     // -1 can appear in a CXVersion, and has various meanings.
     // Whenever one of the fields is -1, the subsequent fields are also -1 (eg
     // you can't have Minor=-1 and Subminor=4). If all 3 fields are -1, it means
     // "no version", and we return null. Otherwise, we treat a -1 in any field
     // as a 0.
     if (Major < 0) return null;
-    return VersionTriple(
-        Major, Minor < 0 ? 0 : Minor, Subminor < 0 ? 0 : Subminor);
+    return Version(Major, Minor < 0 ? 0 : Minor, Subminor < 0 ? 0 : Subminor);
   }
 
   String string() => '$Major.$Minor.$Subminor';