+++ pkgs/ffigen/CHANGELOG.md
@@ -3,7 +3,7 @@
 - Create a public facing API for ffigen that can be invoked as a library:
   `void generate(Config config)`. Make `Config` an implementatble interface,
   rather than needing to be parsed from yaml.
-- Add a `objc-min-target-version` config option. Setting the minimum target
+- Add a `external-versions` config option. Setting the minimum target
   version will omit APIs from the generated bindings if they were deprecated
   before this version.